Multistage Docker build to reduce container size
This multistage Docker build allows removal of unnecessary programs from
the final image. A builder image is used to compile the necessary
dynamic libraries and the application, which are then copied to the
final image.

The libpostal data directory has also been removed and must be mounted
to /libpostal/libpostal for the REST service to work.